Refactor APIv1 code to v1 folder
