+++ services/server/src/server/common.ts
@@ -1,4 +1,3 @@
-import { BadRequestError } from "../common/errors";
 import {
   InvalidSources,
   Language,
@@ -9,7 +8,6 @@ import {
   Status,
   StringMap,
 } from "@ethereum-sourcify/lib-sourcify";
-import { getAddress, isAddress } from "ethers";
 import logger from "../common/logger";
 import { InternalServerError } from "express-openapi-validator/dist/openapi.validator";
 import { Request, Response, NextFunction } from "express";
@@ -30,33 +28,6 @@ export const safeHandler = <T extends Request = Request>(
       );
     }
   };
-};
-
-export const validateSingleAddress = (address: string): boolean => {
-  if (!isAddress(address)) {
-    throw new BadRequestError(`Invalid address: ${address}`);
-  }
-  return true; // if it doesn't throw
-};
-
-export const validateAddresses = (addresses: string): boolean => {
-  const addressesArray = addresses.split(",");
-  const invalidAddresses: string[] = [];
-  for (const i in addressesArray) {
-    const address = addressesArray[i];
-    if (!isAddress(address)) {
-      invalidAddresses.push(address);
-    } else {
-      addressesArray[i] = getAddress(address);
-    }
-  }
-
-  if (invalidAddresses.length) {
-    throw new BadRequestError(
-      `Invalid addresses: ${invalidAddresses.join(", ")}`,
-    );
-  }
-  return true; // if it doesn't throw
 };
 
 export interface PathContentMap {
